Update target SDK and AppCompat to Android Lollipop.
Had a huge issue with AppCompat 21 on KitKat using the dark theme, because the spinner dropdown menu background is white and not black as it is on Lollipop.
I worked around this by inversing the text color in dark themes.
Note that the spinner background arrow drawable is white on pre-Lollipop (even in lite theme) and looks kind of ugly, but I don't see any way around that as it is an issue in AppCompat.