Enable FLAG_NUMBERS for the DateSpinner's weekdays.
When FLAG_NUMBERS and FLAG_WEEKDAY_NAMES are used together, the spinner now shows the date as secondary text next to the weekday name.
Note that in rare occasions the secondary date will be cut off under the spinner arrow as the spinner is not wide enough.
This happens because the PickerSpinnerAdapter cannot measure temporary items when the spinner size is evaluated.