Fix ClassCastException in getSelectedDate (#19).
This is more like a workaround so the onItemSelected callback will now be ignored when the footer is selected.
I also moved the last selected date checks into the date and time spinners, so that redundant onItemSelected calls are ignored when one of the spinners is used without the full ReminderDatePicker.